SW-4761 Move species view into scenes directory
