+++ src/scenes/Species/CheckDataModal.tsx
@@ -3,10 +3,10 @@ import { makeStyles } from '@mui/styles';
 import React, { useEffect, useState } from 'react';
 import strings from 'src/strings';
 import { Species } from 'src/types/Species';
-import Button from '../common/button/Button';
-import DialogBox from '../common/DialogBox/DialogBox';
-import Icon from '../common/icon/Icon';
-import ProgressCircle from '../common/ProgressCircle/ProgressCircle';
+import Button from '../../components/common/button/Button';
+import DialogBox from '../../components/common/DialogBox/DialogBox';
+import Icon from '../../components/common/icon/Icon';
+import ProgressCircle from '../../components/common/ProgressCircle/ProgressCircle';
 
 const useStyles = makeStyles((theme: Theme) => ({
   spacing: {
