+++ src/scenes/Species/ImportSpeciesModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SpeciesService } from 'src/services';
 import strings from 'src/strings';
-import ImportModal from '../common/ImportModal';
+import ImportModal from '../../components/common/ImportModal';
 
 export type ImportSpeciesModalProps = {
   open: boolean;
