+++ src/scenes/Species/TableCellRenderer.tsx
@@ -2,9 +2,9 @@ import { ClickAwayListener, IconButton, Theme, Tooltip } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import React, { useState } from 'react';
 import { SpeciesProblemElement } from 'src/types/Species';
-import Icon from '../common/icon/Icon';
-import CellRenderer, { TableRowType } from '../common/table/TableCellRenderer';
-import { RendererProps } from '../common/table/types';
+import Icon from '../../components/common/icon/Icon';
+import CellRenderer, { TableRowType } from '../../components/common/table/TableCellRenderer';
+import { RendererProps } from '../../components/common/table/types';
 import ProblemTooltip from './ProblemTooltip';
 import { TextTruncated } from '@terraware/web-components';
 import strings from 'src/strings';
